ukssc_telebot-main/requirements.txt
python-telegram-bot==13.9
emoji==1.6.1
backports.zoneinfo==0.2.1
apscheduler==3.6.3
requests==2.26.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
tornado==6.1
pytz==2021.3
tzdata==2021.5
pytz-deprecation-shim==0.1.0.post0
charset-normalizer==2.0.6
tzlocal==4.1
urllib3==1.26.7
pytelegrambotapi==4.1.0
